file handling completed

+++ express_practice/controllers/admin.js
@@ -2,6 +2,7 @@ const { validationResult } = require('express-validator');
 const mongoDb = require('mongodb');
 
 const Product = require('../models/product');
+const fileHelper = require('../utils/file');
 
 exports.getAddProducts = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -16,10 +17,25 @@ exports.getAddProducts = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const image = req.body.image;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
 
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/edit-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description,
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
@@ -29,18 +45,19 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                image: image,
                 price: price,
                 description: description,
-                errorMessage: errors.array()[0].msg,
-                validationErrors: errors.array()
-            }
+            },
+            errorMessage: errors.array()[0].msg,
+            validationErrors: errors.array()
         });
     }
 
+    const imageUrl = image.path;
+
     const product = new Product({
         title: title,
-        image: image,
+        image: imageUrl,
         price: price,
         description: description,
         userId: req.user
@@ -50,7 +67,6 @@ exports.postAddProduct = (req, res, next) => {
     .then(result => {
         res.redirect('/admin/products');
     }).catch(err => {
-        res.redirect('/500');
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
@@ -79,7 +95,6 @@ exports.getEditProducts = (req, res, next) => {
         });
     })
     .catch(err => {
-        res.redirect('/500');
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
@@ -98,7 +113,6 @@ exports.getProducts = (req, res, next) => {
         });
     })
     .catch(err => {
-        res.redirect('/500');
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
@@ -108,7 +122,7 @@ exports.getProducts = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const title = req.body.title;
-    const image = req.body.image;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
 
@@ -121,7 +135,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                image: image,
                 price: price,
                 description: description,
                 _id: prodId
@@ -137,7 +150,10 @@ exports.postEditProduct = (req, res, next) => {
             return res.redirect('/');
         }
         product.title = title;
-        product.image = image;
+        if (image) {
+            fileHelper.deleteFile(product.image);
+            product.image = image.path;
+        }
         product.price = price;
         product.description = description;
         return product.save().then(result => {
@@ -146,7 +162,6 @@ exports.postEditProduct = (req, res, next) => {
         })
     })
     .catch(err => {
-        res.redirect('/500');
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
@@ -155,12 +170,16 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId).then(product => {
+        if (!product) {
+            return new Error('Product not found.')
+        }
+        fileHelper.deleteFile(product.image);
+        return Product.deleteOne({_id: prodId, userId: req.user._id});
+    })
     .then(() => {
-        console.log('Product deleted!');
         res.redirect('/admin/products');
     }).catch(err => {
-        res.redirect('/500');
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
